# ICICI_Server/doc/swagger.json
{
    "info": {
        "version": "v0.0.1",
        "title": "ICICI",
        "description": "To perform any operation, you will need to first generate an access token. \n Tokens are generated by invoking the `/auth` API. \n The response is as follows, `Bearer <access-token>`. \n Copy the response and append it to all subsequent APIs where `Authorization` is required."
    },
...
    "paths": {
...
        "/api/service/file": {
            "get": {
                "summary": "This endpoint will get Image.",
                "description": "Flows imbue Sequence Diagrams.  A sequence diagram shows object interactions arranged in time sequence. It depicts the objects and classes involved in the scenario and the sequence of messages exchanged between the objects needed to carry out the functionality of the scenario. Sequence diagrams are typically associated with use case realizations in the logical-view of the system under development. Sequence diagrams are sometimes called event diagrams or event scenarios. A sequence diagram shows, as parallel vertical lines (lifelines), different processes or objects that live simultaneously, and, as horizontal arrows, the messages exchanged between them, in the order in which they occur. This allows the specification of simple runtime scenarios in a graphical manner. Multiple flows can be added to a Project but only one flow can be selected per Project or Product. A Flow can be navigated through Project and Product.\nThis endpoint will get Image.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/getImage"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "flowId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Flow"
                ]
            }
        },
...
        "/api/file/": {
            "post": {
                "summary": "This endpoint will upload file.",
                "description": "This endpoint will upload file. After that, will read the file using document-reader npm module and store file object in database. ",
                "operationId": "uploadFile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/uploadFile"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "orgName",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "OtherFiles",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "directory",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Document"
                ]
            },
            "get": {
                "summary": "This endpoint will read file data.",
                "description": "This will get file object from database for mapping.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/getFileData"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "flowId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Document"
                ]
            }
        },
        "/api/file/get": {
            "get": {
                "summary": "This endpoint will read file data.",
                "description": "This will get file object from database for mapping.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/getFileData"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Document"
                ]
            }
        },
...
        "/api/file/yaml": {
            "post": {
                "summary": "This endpoint will create YAML file.",
                "description": "This will upload file and creates a yaml for api-connect devlopment.",
                "operationId": "uploadFile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/uploadFile"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "params",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "fileName",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "orgName",
                        "in": "query"
                    }
                ],
                "tags": [
                    "YAML Creation"
                ]
            }
        },
        "/api/file/ui/yaml": {
            "post": {
                "summary": "This endpoint will create YAML file.",
                "description": "This will upload file and creates a yaml for api-connect devlopment.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/uploadFile"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/createYAML"
                        },
                        "required": true,
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "YAML Creation"
                ]
            }
        },
        "/api/file/esql": {
            "post": {
                "summary": "This endpoint will create ESQL file.",
                "description": "This will uplaod file and creates a eqsl for ACE devlopment.",
                "operationId": "uploadFile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/uploadFile"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "mappedObj",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "fileName",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "clientName",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "username",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "description": "",
                        "name": "orgName",
                        "in": "query"
                    }
                ],
                "tags": [
                    "ESQL Creation"
                ]
            }
        },
        "/api/file/ui/esql": {
            "post": {
                "summary": "This endpoint will create ESQL file.",
                "description": "This will uplaod file and creates a eqsl for ACE devlopment.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/uploadFile"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/createESQL"
                        },
                        "required": true,
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "ESQL Creation"
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "description": "Authorization For Transaction APIs",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "createESQL": {
            "properties": {
                "mappedObj": {
                    "type": "string",
                    "description": "Enter mappedObj.",
                    "format": "string"
                },
                "fileName": {
                    "type": "string",
                    "description": "Enter fileName.",
                    "format": "string"
                },
                "clientName": {
                    "type": "string",
                    "description": "Enter clientName.",
                    "format": "string"
                },
                "clientCode": {
                    "type": "string",
                    "description": "Enter clientCode.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter orgName.",
                    "format": "string"
                },
                "productName": {
                    "type": "string",
                    "description": "Enter productName.",
                    "format": "string"
                },
                "serviceName": {
                    "type": "string",
                    "description": "Enter serviceName.",
                    "format": "string"
                },
                "projectId": {
                    "type": "string",
                    "description": "Enter projectId.",
                    "format": "string"
                },
                "fileCount": {
                    "type": "number",
                    "description": "Enter file count.",
                    "format": "string"
                }
            }
        },
        "createYAML": {
            "properties": {
                "params": {
                    "type": "string",
                    "description": "Enter params.",
                    "format": "string"
                },
                "templateName": {
                    "type": "string",
                    "description": "Enter template's File Name.",
                    "format": "string"
                },
                "fileName": {
                    "type": "string",
                    "description": "Enter fileName.",
                    "format": "string"
                },
                "projectId": {
                    "type": "string",
                    "description": "Enter projectId.",
                    "format": "string"
                },
                "username": {
                    "type": "string",
                    "description": "Enter username.",
                    "format": "string"
                },
                "orgName": {
                    "type": "string",
                    "description": "Enter orgName.",
                    "format": "string"
                }
            }
        },
...
    },
    "responses": {
...
        "getImage": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
...
        "getFileData": {
            "required": [],
            "properties": {
                "json": {
                    "type": "string",
                    "description": "You will get json.",
                    "format": "string"
                }
            }
        },
...
    },
...
    "security": [
        {
            "access_token": []
        }
    ],
    "swagger": "2.0"
}